+++ datanode/gateway/graphql/generated.go
@@ -150,6 +150,7 @@ type ResolverRoot interface {
 	TradeUpdate() TradeUpdateResolver
 	TransactionResult() TransactionResultResolver
 	Transfer() TransferResolver
+	TransferNode() TransferNodeResolver
 	UpdateAsset() UpdateAssetResolver
 	UpdateMarket() UpdateMarketResolver
 	UpdateMarketConfiguration() UpdateMarketConfigurationResolver
@@ -2434,6 +2435,17 @@ type ComplexityRoot struct {
 		Node   func(childComplexity int) int
 	}
 
+	TransferFee struct {
+		Amount     func(childComplexity int) int
+		Epoch      func(childComplexity int) int
+		TransferID func(childComplexity int) int
+	}
+
+	TransferNode struct {
+		Fees     func(childComplexity int) int
+		Transfer func(childComplexity int) int
+	}
+
 	TransferResponse struct {
 		Balances  func(childComplexity int) int
 		Transfers func(childComplexity int) int
@@ -3350,6 +3362,9 @@ type TransferResolver interface {
 	Asset(ctx context.Context, obj *v1.Transfer) (*vega.Asset, error)
 
 	Kind(ctx context.Context, obj *v1.Transfer) (TransferKind, error)
+}
+type TransferNodeResolver interface {
+	Fees(ctx context.Context, obj *v2.TransferNode) ([]*TransferFee, error)
 }
 type UpdateAssetResolver interface {
 	Quantum(ctx context.Context, obj *vega.UpdateAsset) (string, error)
@@ -13508,6 +13523,41 @@ func (e *executableSchema) Complexity(typeName, field string, childComplexity in
 		}
 
 		return e.complexity.TransferEdge.Node(childComplexity), true
+
+	case "TransferFee.amount":
+		if e.complexity.TransferFee.Amount == nil {
+			break
+		}
+
+		return e.complexity.TransferFee.Amount(childComplexity), true
+
+	case "TransferFee.epoch":
+		if e.complexity.TransferFee.Epoch == nil {
+			break
+		}
+
+		return e.complexity.TransferFee.Epoch(childComplexity), true
+
+	case "TransferFee.transferId":
+		if e.complexity.TransferFee.TransferID == nil {
+			break
+		}
+
+		return e.complexity.TransferFee.TransferID(childComplexity), true
+
+	case "TransferNode.fees":
+		if e.complexity.TransferNode.Fees == nil {
+			break
+		}
+
+		return e.complexity.TransferNode.Fees(childComplexity), true
+
+	case "TransferNode.transfer":
+		if e.complexity.TransferNode.Transfer == nil {
+			break
+		}
+
+		return e.complexity.TransferNode.Transfer(childComplexity), true
 
 	case "TransferResponse.balances":
 		if e.complexity.TransferResponse.Balances == nil {
@@ -84598,9 +84648,9 @@ func (ec *executionContext) _TransferEdge_node(ctx context.Context, field graphq
 		}
 		return graphql.Null
 	}
-	res := resTmp.(*v1.Transfer)
+	res := resTmp.(*v2.TransferNode)
 	fc.Result = res
-	return ec.marshalNTransfer2ᚖcodeᚗvegaprotocolᚗioᚋvegaᚋprotosᚋvegaᚋeventsᚋv1ᚐTransfer(ctx, field.Selections, res)
+	return ec.marshalNTransferNode2ᚖcodeᚗvegaprotocolᚗioᚋvegaᚋprotosᚋdataᚑnodeᚋapiᚋv2ᚐTransferNode(ctx, field.Selections, res)
 }
 
 func (ec *executionContext) fieldContext_TransferEdge_node(ctx context.Context, field graphql.CollectedField) (fc *graphql.FieldContext, err error) {
@@ -84611,32 +84661,12 @@ func (ec *executionContext) fieldContext_TransferEdge_node(ctx context.Context, 
 		IsResolver: false,
 		Child: func(ctx context.Context, field graphql.CollectedField) (*graphql.FieldContext, error) {
 			switch field.Name {
-			case "id":
-				return ec.fieldContext_Transfer_id(ctx, field)
-			case "from":
-				return ec.fieldContext_Transfer_from(ctx, field)
-			case "fromAccountType":
-				return ec.fieldContext_Transfer_fromAccountType(ctx, field)
-			case "to":
-				return ec.fieldContext_Transfer_to(ctx, field)
-			case "toAccountType":
-				return ec.fieldContext_Transfer_toAccountType(ctx, field)
-			case "asset":
-				return ec.fieldContext_Transfer_asset(ctx, field)
-			case "amount":
-				return ec.fieldContext_Transfer_amount(ctx, field)
-			case "reference":
-				return ec.fieldContext_Transfer_reference(ctx, field)
-			case "status":
-				return ec.fieldContext_Transfer_status(ctx, field)
-			case "timestamp":
-				return ec.fieldContext_Transfer_timestamp(ctx, field)
-			case "kind":
-				return ec.fieldContext_Transfer_kind(ctx, field)
-			case "reason":
-				return ec.fieldContext_Transfer_reason(ctx, field)
+			case "transfer":
+				return ec.fieldContext_TransferNode_transfer(ctx, field)
+			case "fees":
+				return ec.fieldContext_TransferNode_fees(ctx, field)
 			}
-			return nil, fmt.Errorf("no field named %q was found under type Transfer", field.Name)
+			return nil, fmt.Errorf("no field named %q was found under type TransferNode", field.Name)
 		},
 	}
 	return fc, nil
@@ -84681,6 +84711,257 @@ func (ec *executionContext) fieldContext_TransferEdge_cursor(ctx context.Context
 		IsResolver: false,
 		Child: func(ctx context.Context, field graphql.CollectedField) (*graphql.FieldContext, error) {
 			return nil, errors.New("field of type String does not have child fields")
+		},
+	}
+	return fc, nil
+}
+
+func (ec *executionContext) _TransferFee_transferId(ctx context.Context, field graphql.CollectedField, obj *TransferFee) (ret graphql.Marshaler) {
+	fc, err := ec.fieldContext_TransferFee_transferId(ctx, field)
+	if err != nil {
+		return graphql.Null
+	}
+	ctx = graphql.WithFieldContext(ctx, fc)
+	defer func() {
+		if r := recover(); r != nil {
+			ec.Error(ctx, ec.Recover(ctx, r))
+			ret = graphql.Null
+		}
+	}()
+	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return obj.TransferID, nil
+	})
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	if resTmp == nil {
+		if !graphql.HasFieldError(ctx, fc) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(string)
+	fc.Result = res
+	return ec.marshalNID2string(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) fieldContext_TransferFee_transferId(ctx context.Context, field graphql.CollectedField) (fc *graphql.FieldContext, err error) {
+	fc = &graphql.FieldContext{
+		Object:     "TransferFee",
+		Field:      field,
+		IsMethod:   false,
+		IsResolver: false,
+		Child: func(ctx context.Context, field graphql.CollectedField) (*graphql.FieldContext, error) {
+			return nil, errors.New("field of type ID does not have child fields")
+		},
+	}
+	return fc, nil
+}
+
+func (ec *executionContext) _TransferFee_amount(ctx context.Context, field graphql.CollectedField, obj *TransferFee) (ret graphql.Marshaler) {
+	fc, err := ec.fieldContext_TransferFee_amount(ctx, field)
+	if err != nil {
+		return graphql.Null
+	}
+	ctx = graphql.WithFieldContext(ctx, fc)
+	defer func() {
+		if r := recover(); r != nil {
+			ec.Error(ctx, ec.Recover(ctx, r))
+			ret = graphql.Null
+		}
+	}()
+	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return obj.Amount, nil
+	})
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	if resTmp == nil {
+		if !graphql.HasFieldError(ctx, fc) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(string)
+	fc.Result = res
+	return ec.marshalNString2string(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) fieldContext_TransferFee_amount(ctx context.Context, field graphql.CollectedField) (fc *graphql.FieldContext, err error) {
+	fc = &graphql.FieldContext{
+		Object:     "TransferFee",
+		Field:      field,
+		IsMethod:   false,
+		IsResolver: false,
+		Child: func(ctx context.Context, field graphql.CollectedField) (*graphql.FieldContext, error) {
+			return nil, errors.New("field of type String does not have child fields")
+		},
+	}
+	return fc, nil
+}
+
+func (ec *executionContext) _TransferFee_epoch(ctx context.Context, field graphql.CollectedField, obj *TransferFee) (ret graphql.Marshaler) {
+	fc, err := ec.fieldContext_TransferFee_epoch(ctx, field)
+	if err != nil {
+		return graphql.Null
+	}
+	ctx = graphql.WithFieldContext(ctx, fc)
+	defer func() {
+		if r := recover(); r != nil {
+			ec.Error(ctx, ec.Recover(ctx, r))
+			ret = graphql.Null
+		}
+	}()
+	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return obj.Epoch, nil
+	})
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	if resTmp == nil {
+		if !graphql.HasFieldError(ctx, fc) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(int)
+	fc.Result = res
+	return ec.marshalNInt2int(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) fieldContext_TransferFee_epoch(ctx context.Context, field graphql.CollectedField) (fc *graphql.FieldContext, err error) {
+	fc = &graphql.FieldContext{
+		Object:     "TransferFee",
+		Field:      field,
+		IsMethod:   false,
+		IsResolver: false,
+		Child: func(ctx context.Context, field graphql.CollectedField) (*graphql.FieldContext, error) {
+			return nil, errors.New("field of type Int does not have child fields")
+		},
+	}
+	return fc, nil
+}
+
+func (ec *executionContext) _TransferNode_transfer(ctx context.Context, field graphql.CollectedField, obj *v2.TransferNode) (ret graphql.Marshaler) {
+	fc, err := ec.fieldContext_TransferNode_transfer(ctx, field)
+	if err != nil {
+		return graphql.Null
+	}
+	ctx = graphql.WithFieldContext(ctx, fc)
+	defer func() {
+		if r := recover(); r != nil {
+			ec.Error(ctx, ec.Recover(ctx, r))
+			ret = graphql.Null
+		}
+	}()
+	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return obj.Transfer, nil
+	})
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	if resTmp == nil {
+		if !graphql.HasFieldError(ctx, fc) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(*v1.Transfer)
+	fc.Result = res
+	return ec.marshalNTransfer2ᚖcodeᚗvegaprotocolᚗioᚋvegaᚋprotosᚋvegaᚋeventsᚋv1ᚐTransfer(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) fieldContext_TransferNode_transfer(ctx context.Context, field graphql.CollectedField) (fc *graphql.FieldContext, err error) {
+	fc = &graphql.FieldContext{
+		Object:     "TransferNode",
+		Field:      field,
+		IsMethod:   false,
+		IsResolver: false,
+		Child: func(ctx context.Context, field graphql.CollectedField) (*graphql.FieldContext, error) {
+			switch field.Name {
+			case "id":
+				return ec.fieldContext_Transfer_id(ctx, field)
+			case "from":
+				return ec.fieldContext_Transfer_from(ctx, field)
+			case "fromAccountType":
+				return ec.fieldContext_Transfer_fromAccountType(ctx, field)
+			case "to":
+				return ec.fieldContext_Transfer_to(ctx, field)
+			case "toAccountType":
+				return ec.fieldContext_Transfer_toAccountType(ctx, field)
+			case "asset":
+				return ec.fieldContext_Transfer_asset(ctx, field)
+			case "amount":
+				return ec.fieldContext_Transfer_amount(ctx, field)
+			case "reference":
+				return ec.fieldContext_Transfer_reference(ctx, field)
+			case "status":
+				return ec.fieldContext_Transfer_status(ctx, field)
+			case "timestamp":
+				return ec.fieldContext_Transfer_timestamp(ctx, field)
+			case "kind":
+				return ec.fieldContext_Transfer_kind(ctx, field)
+			case "reason":
+				return ec.fieldContext_Transfer_reason(ctx, field)
+			}
+			return nil, fmt.Errorf("no field named %q was found under type Transfer", field.Name)
+		},
+	}
+	return fc, nil
+}
+
+func (ec *executionContext) _TransferNode_fees(ctx context.Context, field graphql.CollectedField, obj *v2.TransferNode) (ret graphql.Marshaler) {
+	fc, err := ec.fieldContext_TransferNode_fees(ctx, field)
+	if err != nil {
+		return graphql.Null
+	}
+	ctx = graphql.WithFieldContext(ctx, fc)
+	defer func() {
+		if r := recover(); r != nil {
+			ec.Error(ctx, ec.Recover(ctx, r))
+			ret = graphql.Null
+		}
+	}()
+	resTmp, err := ec.ResolverMiddleware(ctx, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return ec.resolvers.TransferNode().Fees(rctx, obj)
+	})
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	if resTmp == nil {
+		return graphql.Null
+	}
+	res := resTmp.([]*TransferFee)
+	fc.Result = res
+	return ec.marshalOTransferFee2ᚕᚖcodeᚗvegaprotocolᚗioᚋvegaᚋdatanodeᚋgatewayᚋgraphqlᚐTransferFee(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) fieldContext_TransferNode_fees(ctx context.Context, field graphql.CollectedField) (fc *graphql.FieldContext, err error) {
+	fc = &graphql.FieldContext{
+		Object:     "TransferNode",
+		Field:      field,
+		IsMethod:   true,
+		IsResolver: true,
+		Child: func(ctx context.Context, field graphql.CollectedField) (*graphql.FieldContext, error) {
+			switch field.Name {
+			case "transferId":
+				return ec.fieldContext_TransferFee_transferId(ctx, field)
+			case "amount":
+				return ec.fieldContext_TransferFee_amount(ctx, field)
+			case "epoch":
+				return ec.fieldContext_TransferFee_epoch(ctx, field)
+			}
+			return nil, fmt.Errorf("no field named %q was found under type TransferFee", field.Name)
 		},
 	}
 	return fc, nil
@@ -112800,6 +113081,93 @@ func (ec *executionContext) _TransferEdge(ctx context.Context, sel ast.Selection
 	return out
 }
 
+var transferFeeImplementors = []string{"TransferFee"}
+
+func (ec *executionContext) _TransferFee(ctx context.Context, sel ast.SelectionSet, obj *TransferFee) graphql.Marshaler {
+	fields := graphql.CollectFields(ec.OperationContext, sel, transferFeeImplementors)
+	out := graphql.NewFieldSet(fields)
+	var invalids uint32
+	for i, field := range fields {
+		switch field.Name {
+		case "__typename":
+			out.Values[i] = graphql.MarshalString("TransferFee")
+		case "transferId":
+
+			out.Values[i] = ec._TransferFee_transferId(ctx, field, obj)
+
+			if out.Values[i] == graphql.Null {
+				invalids++
+			}
+		case "amount":
+
+			out.Values[i] = ec._TransferFee_amount(ctx, field, obj)
+
+			if out.Values[i] == graphql.Null {
+				invalids++
+			}
+		case "epoch":
+
+			out.Values[i] = ec._TransferFee_epoch(ctx, field, obj)
+
+			if out.Values[i] == graphql.Null {
+				invalids++
+			}
+		default:
+			panic("unknown field " + strconv.Quote(field.Name))
+		}
+	}
+	out.Dispatch()
+	if invalids > 0 {
+		return graphql.Null
+	}
+	return out
+}
+
+var transferNodeImplementors = []string{"TransferNode"}
+
+func (ec *executionContext) _TransferNode(ctx context.Context, sel ast.SelectionSet, obj *v2.TransferNode) graphql.Marshaler {
+	fields := graphql.CollectFields(ec.OperationContext, sel, transferNodeImplementors)
+	out := graphql.NewFieldSet(fields)
+	var invalids uint32
+	for i, field := range fields {
+		switch field.Name {
+		case "__typename":
+			out.Values[i] = graphql.MarshalString("TransferNode")
+		case "transfer":
+
+			out.Values[i] = ec._TransferNode_transfer(ctx, field, obj)
+
+			if out.Values[i] == graphql.Null {
+				atomic.AddUint32(&invalids, 1)
+			}
+		case "fees":
+			field := field
+
+			innerFunc := func(ctx context.Context) (res graphql.Marshaler) {
+				defer func() {
+					if r := recover(); r != nil {
+						ec.Error(ctx, ec.Recover(ctx, r))
+					}
+				}()
+				res = ec._TransferNode_fees(ctx, field, obj)
+				return res
+			}
+
+			out.Concurrently(i, func() graphql.Marshaler {
+				return innerFunc(ctx)
+
+			})
+		default:
+			panic("unknown field " + strconv.Quote(field.Name))
+		}
+	}
+	out.Dispatch()
+	if invalids > 0 {
+		return graphql.Null
+	}
+	return out
+}
+
 var transferResponseImplementors = []string{"TransferResponse"}
 
 func (ec *executionContext) _TransferResponse(ctx context.Context, sel ast.SelectionSet, obj *TransferResponse) graphql.Marshaler {
@@ -118862,6 +119230,16 @@ func (ec *executionContext) marshalNTransferKind2codeᚗvegaprotocolᚗioᚋvega
 	return ec._TransferKind(ctx, sel, v)
 }
 
+func (ec *executionContext) marshalNTransferNode2ᚖcodeᚗvegaprotocolᚗioᚋvegaᚋprotosᚋdataᚑnodeᚋapiᚋv2ᚐTransferNode(ctx context.Context, sel ast.SelectionSet, v *v2.TransferNode) graphql.Marshaler {
+	if v == nil {
+		if !graphql.HasFieldError(ctx, graphql.GetFieldContext(ctx)) {
+			ec.Errorf(ctx, "the requested element is null which the schema does not allow")
+		}
+		return graphql.Null
+	}
+	return ec._TransferNode(ctx, sel, v)
+}
+
 func (ec *executionContext) marshalNTransferResponse2ᚖcodeᚗvegaprotocolᚗioᚋvegaᚋdatanodeᚋgatewayᚋgraphqlᚐTransferResponse(ctx context.Context, sel ast.SelectionSet, v *TransferResponse) graphql.Marshaler {
 	if v == nil {
 		if !graphql.HasFieldError(ctx, graphql.GetFieldContext(ctx)) {
@@ -124861,6 +125239,54 @@ func (ec *executionContext) marshalOTransferEdge2ᚖcodeᚗvegaprotocolᚗioᚋv
 		return graphql.Null
 	}
 	return ec._TransferEdge(ctx, sel, v)
+}
+
+func (ec *executionContext) marshalOTransferFee2ᚕᚖcodeᚗvegaprotocolᚗioᚋvegaᚋdatanodeᚋgatewayᚋgraphqlᚐTransferFee(ctx context.Context, sel ast.SelectionSet, v []*TransferFee) graphql.Marshaler {
+	if v == nil {
+		return graphql.Null
+	}
+	ret := make(graphql.Array, len(v))
+	var wg sync.WaitGroup
+	isLen1 := len(v) == 1
+	if !isLen1 {
+		wg.Add(len(v))
+	}
+	for i := range v {
+		i := i
+		fc := &graphql.FieldContext{
+			Index:  &i,
+			Result: &v[i],
+		}
+		ctx := graphql.WithFieldContext(ctx, fc)
+		f := func(i int) {
+			defer func() {
+				if r := recover(); r != nil {
+					ec.Error(ctx, ec.Recover(ctx, r))
+					ret = nil
+				}
+			}()
+			if !isLen1 {
+				defer wg.Done()
+			}
+			ret[i] = ec.marshalOTransferFee2ᚖcodeᚗvegaprotocolᚗioᚋvegaᚋdatanodeᚋgatewayᚋgraphqlᚐTransferFee(ctx, sel, v[i])
+		}
+		if isLen1 {
+			f(i)
+		} else {
+			go f(i)
+		}
+
+	}
+	wg.Wait()
+
+	return ret
+}
+
+func (ec *executionContext) marshalOTransferFee2ᚖcodeᚗvegaprotocolᚗioᚋvegaᚋdatanodeᚋgatewayᚋgraphqlᚐTransferFee(ctx context.Context, sel ast.SelectionSet, v *TransferFee) graphql.Marshaler {
+	if v == nil {
+		return graphql.Null
+	}
+	return ec._TransferFee(ctx, sel, v)
 }
 
 func (ec *executionContext) marshalOTransferResponse2ᚕᚖcodeᚗvegaprotocolᚗioᚋvegaᚋdatanodeᚋgatewayᚋgraphqlᚐTransferResponseᚄ(ctx context.Context, sel ast.SelectionSet, v []*TransferResponse) graphql.Marshaler {