+++ datanode/sqlsubscribers/transfers.go
@@ -42,8 +42,14 @@ type TransferEvent interface {
 	TransferFunds() eventspb.Transfer
 }
 
+type TransferFeesEvent interface {
+	events.Event
+	TransferFees() eventspb.TransferFees
+}
+
 type TransferStore interface {
 	Upsert(ctx context.Context, transfer *entities.Transfer) error
+	UpsertFees(ctx context.Context, tf *entities.TransferFees) error
 }
 
 type AccountSource interface {
@@ -65,11 +71,20 @@ func NewTransfer(store TransferStore, accountSource AccountSource) *Transfer {
 }
 
 func (rf *Transfer) Types() []events.Type {
-	return []events.Type{events.TransferEvent}
+	return []events.Type{
+		events.TransferEvent,
+		events.TransferFeesEvent,
+	}
 }
 
 func (rf *Transfer) Push(ctx context.Context, evt events.Event) error {
-	return rf.consume(ctx, evt.(TransferEvent))
+	switch te := evt.(type) {
+	case TransferEvent:
+		return rf.consume(ctx, te)
+	case TransferFeesEvent:
+		return rf.handleFees(ctx, te)
+	}
+	return errors.New("unsupported event")
 }
 
 func (rf *Transfer) consume(ctx context.Context, event TransferEvent) error {
@@ -80,4 +95,10 @@ func (rf *Transfer) consume(ctx context.Context, event TransferEvent) error {
 	}
 
 	return errors.Wrap(rf.store.Upsert(ctx, record), "inserting transfer into to SQL store failed")
+}
+
+func (rf *Transfer) handleFees(ctx context.Context, e TransferFeesEvent) error {
+	tf := e.TransferFees()
+	rec := entities.TransferFeesFromProto(&tf, rf.vegaTime)
+	return errors.Wrap(rf.store.UpsertFees(ctx, rec), "inserting transfer fee into SQL store failed")
 }